# MAINT: Optimize the logical implementation for RISC-V based on Highway Wrapper