Test environment: Banana Pi BPI-F3 (SpacemiT K1 8-core RISC-V chip) gcc version 14.2.0
Test command: spin bench --compare HEAD~1 -t bench_ufunc.LogicalSpecific

Add class LogicalSpecific in benchmarks/benchmarks/bench_ufunc.py for this test on Banana Pi BPI-F3

```
class LogicalSpecific(Benchmark):
    timeout = 10
    params = ['bool']
    param_names = ['dtype']

    def setup(self, dtype):
        np.seterr(all='ignore')
        self.logical_and_fn = np.logical_and
        self.logical_or_fn = np.logical_or
        self.logical_not_fn = np.logical_not
        self.absolute_fn = np.absolute

        from .common import get_square_
        arr1 = get_square_(dtype)
        arr2 = get_square_(dtype)

        self.data_and = (arr1, arr2)
        self.data_or  = (arr1, arr2)
        self.data_not = (arr1,)
        self.data_abs = (arr1,)

    def time_logical_and(self, dtype):
        self.logical_and_fn(*self.data_and)

    def time_logical_or(self, dtype):
        self.logical_or_fn(*self.data_or)

    def time_logical_not(self, dtype):
        self.logical_not_fn(*self.data_not)

    def time_absolute(self, dtype):
        self.absolute_fn(*self.data_abs)

```
(numpy_venv) root@yangwang:/home/yangwang/numpy# git log -2
commit e5446650f2f29ac2c21e459835d9b04baa1b131d (HEAD -> main)
Author: Wang Yang <yangwang@iscas.ac.cn>
Date:   Fri Aug 29 15:39:43 2025 +0800

ENH, SIMD: Optimize the logical implementation based on Highway Wrapper

commit b112b4e4cc12ae8a0968b26dcaa53ba032fced29
Author: Wang Yang <yangwang@iscas.ac.cn>
Date:   Fri Aug 29 15:38:43 2025 +0800

TST:Add logical test

(numpy_venv) root@yangwang:/home/yangwang/numpy#
(numpy_venv) root@yangwang:/home/yangwang/numpy# spin bench --compare HEAD~1 -t bench_ufunc.LogicalSpecific
| Change   | Before [b112b4e4] <main~1>   | After [e5446650] <main>   |   Ratio | Benchmark (Parameter)                                |
|----------|------------------------------|---------------------------|---------|------------------------------------------------------|
| -        | 92.2±0.5μs                   | 43.4±8μs                  |    0.47 | bench_ufunc.LogicalSpecific.time_logical_not('bool') |
| -        | 90.3±0.9μs                   | 41.9±8μs                  |    0.46 | bench_ufunc.LogicalSpecific.time_absolute('bool')    |
| -        | 128±0.3μs                    | 56.8±10μs                 |    0.44 | bench_ufunc.LogicalSpecific.time_logical_or('bool')  |
| -        | 127±1μs                      | 55.1±10μs                 |    0.43 | bench_ufunc.LogicalSpecific.time_logical_and('bool') |