@mattip 2025-09-05
> Add class LogicalSpecific 

Could you add the benchmark? Preferably in a separate PR so it can be part of the code base and enable comparison to this PR